refactor: riorganizza CLI inspect in 3 subcomandi (config, summary, runs)
- inspect/config: unifica layer + schema-diff --diff + profile + query
- inspect/summary: unifica status
- inspect/runs: cronologia run (lista + dettaglio) + resume
- layer/status/resume diventano hidden alias che delegano ai nuovi inspect
- inspect profile/schema/paths/schema-diff/sparql restano hidden
- Tutti i backend condivisi con MCP (layer_ops, readiness_ops, ...)
- 1088 test passano, 0 regressions
